shuffle around tests
